+++ pkgs/claude_box/lib/claude_box.dart
@@ -1,5 +1,6 @@
 export 'package:claude_box/src/claude.dart';
 export 'package:claude_box/src/claude_core.dart';
+export 'package:claude_box/src/claude_stream.dart';
 export 'package:claude_box/src/freezed/message/request/message_request/message_request.dart';
 export 'package:claude_box/src/freezed/message/response/message_response.dart';
 export 'package:claude_box/src/freezed/models/list_models_query_parameters/list_models_query_parameters.dart';